Customer_Segmentation
The crux of this repository is to realize that the most important part about any analytics or Machine Learning-based solution is the value it can deliver to the organization. We will look at ideas pertaining to customer segmentation, its impact, and explore a novel way of using unsupervised learning to find out customer segments and view interesting patterns and behavior.